+++ src/main/java/com/yuuko/core/commands/core/settings/SettingExecuteBoolean.java
@@ -5,9 +5,17 @@ import com.yuuko.core.database.function.GuildFunctions;
 import com.yuuko.core.events.entity.MessageEvent;
 import net.dv8tion.jda.api.EmbedBuilder;
 
-import java.awt.*;
+import java.util.Arrays;
+import java.util.List;
 
 public class SettingExecuteBoolean extends Setting {
+
+    private static final List<String> booleanValues = Arrays.asList(
+            "true",
+            "yes",
+            "false",
+            "no"
+    );
 
     public SettingExecuteBoolean(MessageEvent e) {
         onCommand(e);
@@ -17,17 +25,25 @@ public class SettingExecuteBoolean extends Setting {
         try {
             String[] parameters = e.getParameters().toLowerCase().split("\\s+", 2);
 
-            if(parameters.length > 1) {
-                String intValue = (parameters[1].equals("true") || parameters[1].equals("yes")) ? "1" : "0";
+            if(parameters.length < 2) {
+                return;
+            }
 
-                if(GuildFunctions.setGuildSettings(parameters[0], intValue, e.getGuild().getId())) {
-                    if(Boolean.parseBoolean(parameters[1].toUpperCase())) {
-                        EmbedBuilder embed = new EmbedBuilder().setColor(Color.GREEN).setTitle(parameters[0].toUpperCase() + " set to TRUE.");
-                        MessageHandler.sendMessage(e, embed.build());
-                    } else {
-                        EmbedBuilder embed = new EmbedBuilder().setColor(Color.RED).setTitle(parameters[0].toUpperCase() + " set to FALSE.");
-                        MessageHandler.sendMessage(e, embed.build());
-                    }
+            if(!booleanValues.contains(parameters[1])) {
+                EmbedBuilder embed = new EmbedBuilder().setTitle("_" + parameters[1].toUpperCase() + "_ is not a valid value. (Valid: TRUE, FALSE)");
+                MessageHandler.sendMessage(e, embed.build());
+                return;
+            }
+
+            String intValue = (parameters[1].equals("true") || parameters[1].equals("yes")) ? "1" : "0";
+
+            if(GuildFunctions.setGuildSettings(parameters[0], intValue, e.getGuild().getId())) {
+                if(Boolean.parseBoolean(parameters[1].toUpperCase())) {
+                    EmbedBuilder embed = new EmbedBuilder().setTitle(parameters[0].toUpperCase() + " set to TRUE.");
+                    MessageHandler.sendMessage(e, embed.build());
+                } else {
+                    EmbedBuilder embed = new EmbedBuilder().setTitle(parameters[0].toUpperCase() + " set to FALSE.");
+                    MessageHandler.sendMessage(e, embed.build());
                 }
             }
         } catch(Exception ex) {